doc(Examples): added new example for httpsClient
